Test App for the SPI RX function - Commands between SMT Master and Arduino Slave - Unfinished
First command to control an Arduino port (LED) is implemented.
Basic comuunication protocol with a very basic errors detection/recovery only.
Open issues:
Delay is needed in line 190, else the program does not work.
If the SPI clock speed is increased to 8MHz instead of 2MHz, the data is corrupted even with this delay - need to be debugged.